+++ packages/docusaurus-plugin-content-docs/src/__tests__/lastUpdate.test.ts
@@ -6,10 +6,8 @@
  */
 
 import {jest} from '@jest/globals';
-import fs from 'fs-extra';
 import path from 'path';
 import shell from 'shelljs';
-import {createTempRepo} from '@testing-utils/git';
 import {getFileLastUpdate} from '@docusaurus/utils';
 
 describe('getFileLastUpdate', () => {
@@ -60,40 +58,6 @@ describe('getFileLastUpdate', () => {
     );
     expect(consoleMock).toHaveBeenCalledTimes(0);
     consoleMock.mockRestore();
-  });
-
-  it('temporary created file that is not tracked by git', async () => {
-    const consoleMock = jest
-      .spyOn(console, 'warn')
-      .mockImplementation(() => {});
-    const {repoDir} = createTempRepo();
-    const tempFilePath = path.join(repoDir, 'file.md');
-    await fs.writeFile(tempFilePath, 'Lorem ipsum :)');
-    await expect(getFileLastUpdate(tempFilePath)).resolves.toBeNull();
-    expect(consoleMock).toHaveBeenCalledTimes(1);
-    expect(consoleMock).toHaveBeenLastCalledWith(
-      expect.stringMatching(/not tracked by git./),
-    );
-    await fs.unlink(tempFilePath);
-  });
-
-  it('multiple files not tracked by git', async () => {
-    const consoleMock = jest
-      .spyOn(console, 'warn')
-      .mockImplementation(() => {});
-    const {repoDir} = createTempRepo();
-    const tempFilePath1 = path.join(repoDir, 'file1.md');
-    const tempFilePath2 = path.join(repoDir, 'file2.md');
-    await fs.writeFile(tempFilePath1, 'Lorem ipsum :)');
-    await fs.writeFile(tempFilePath2, 'Lorem ipsum :)');
-    await expect(getFileLastUpdate(tempFilePath1)).resolves.toBeNull();
-    await expect(getFileLastUpdate(tempFilePath2)).resolves.toBeNull();
-    expect(consoleMock).toHaveBeenCalledTimes(1);
-    expect(consoleMock).toHaveBeenLastCalledWith(
-      expect.stringMatching(/not tracked by git./),
-    );
-    await fs.unlink(tempFilePath1);
-    await fs.unlink(tempFilePath2);
   });
 
   it('git does not exist', async () => {
